backend/OIA/requirements.txt
asgiref==3.9.1
defusedxml==0.7.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
numpy==2.3.1
odfpy==1.4.1
openpyxl==3.1.5
pandas==2.3.1
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
xlsxwriter==3.2.5
